+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v1.1.0
+
+- (feature) Add `RunWithCtx()` to `Connection`
+
 # v1.0.0
 
 - (feature) Implement `graphql-transport-ws` subprotocol